feat(aep-api): contract-first tracer bullet — strict server serves orgs + projects
The generated strict server (oapi-codegen v2.7 strict+std-http from
packages/contracts/api/v1) replaces the Huma mux at /api/v1:

- make gen-api: schema types generate INTO models/ (models_gen.go — same
  import path, prefer-skip-optional-pointer + initialisms normalizer keeps
  every importer compiling), strict server into internal/api/gen/, contract
  vendored for embedding; gen-api-check is the CI freshness gate
- 22 hand-written spec-shaped structs deleted from models/ (gorm-tagged
  AccessRequest/ComponentConfig/RcaAgentReport stay via exclude-schemas)
- deny-by-default tenant gate (tenant_gate.go): every operation requires the
  org claim from the verified JWT unless enumerated in tenantGateCarveOuts
  (exactly one entry: ListOrganizations); org is bound into context and
  passed to services explicitly; gate mode stays request-context ENFORCE/LOG
- kin-openapi request validator against the embedded committed contract:
  schema-invalid requests answer 400 {code: validation_failed, details[]}
  before any handler; route-miss falls through (read-file catch-all)
- flat error envelope everywhere on the edge (errors.go): typed *apiError +
  central writers replace the Huma error constructors; 501 not_implemented
  stubs (stubs.go) answer every unmigrated operation until issue 003
- Organizations + Projects implemented on the strict interface
  (handlers_*.go), their *_huma.go deleted, component tests ported (only
  status/envelope assertions changed); /openapi.yaml + /components.yaml now
  serve the committed contract byte-for-byte
- new arch-locks: carve-out set pinned against the generated interface,
  gate unit tests, 501-stub smoke; contract fixes surfaced by real codegen
  ($ref siblings, {} schema, enum var names, x-go-type overrides)

Migration window: component/execution/files/genai/requirements/task
component tests are red against the 501 stubs by design — issue 003 ports
them feature by feature. codingagent/skills failures are pre-existing.

Part of the contract-first backend migration (local issue 002).

+++ services/aep-api/internal/clients/observability/client.go
@@ -112,7 +112,7 @@ func (c *observabilityClient) GetBuildLogs(ctx context.Context, orgName, project
 
 	logs := &models.BuildLogs{}
 	if result.TotalCount != nil {
-		logs.TotalCount = *result.TotalCount
+		logs.TotalCount = int64(*result.TotalCount)
 	}
 	if result.Logs != nil {
 		for _, e := range *result.Logs {
